added logic to Controller, model, repo and services.

+++ CRUD-PRODUCTOS/productos/src/main/java/ecoMarket/duoc/cl/productos/repository/ProductoRepository.java
@@ -3,5 +3,5 @@ package ecoMarket.duoc.cl.productos.repository;
 import ecoMarket.duoc.cl.productos.model.Producto;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductoRepository extends JpaRepository<Producto, Integer> {
+public interface ProductoRepository extends JpaRepository<Producto, Long> {
 }
